# docs/swagger/swagger-sdce-1.json
{
  "openapi" : "3.0.1",
  "info" : {
    "contact" : {
      "email" : "onap-discuss@lists.onap.org",
      "name" : "ONAP",
      "url" : "https://onap.readthedocs.io"
    },
    "description" : "SDC API for onboarding (SDCE-1)",
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0"
    },
    "title" : "SPC API: SDCE-1",
    "version" : "1.0"
  },
  "servers" : [ {
    "description" : "SDCE-1 APIs",
    "url" : "/sdc"
  } ],
  "paths" : {
...
    "/v1.0/vendor-software-products/{vspId}/versions/{versionId}/components/{componentId}/processes/{processId}/upload" : {
      "delete" : {
        "description" : "Delete vendor software product process uploaded file",
        "operationId" : "deleteUploadedFile",
        "parameters" : [ {
          "description" : "Vendor software product Id",
          "in" : "path",
          "name" : "vspId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor software product version Id",
          "in" : "path",
          "name" : "versionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor software product component Id",
          "in" : "path",
          "name" : "componentId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor software product process Id",
          "in" : "path",
          "name" : "processId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "USER_ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/json" : { }
            },
            "description" : "default response"
          }
        },
        "tags" : [ "SDCE-1 APIs", "Vendor Software Product Component Processes" ]
      },
      "get" : {
        "description" : "Get vendor software product process uploaded file",
        "operationId" : "getUploadedFile",
        "parameters" : [ {
          "description" : "Vendor software product Id",
          "in" : "path",
          "name" : "vspId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor software product version Id",
          "in" : "path",
          "name" : "versionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor software product component Id",
          "in" : "path",
          "name" : "componentId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor software product process Id",
          "in" : "path",
          "name" : "processId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "USER_ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/octet-stream" : { }
            },
            "description" : "default response"
          }
        },
        "tags" : [ "SDCE-1 APIs", "Vendor Software Product Component Processes" ]
      },
      "post" : {
        "description" : "Update vendor software product process upload",
        "operationId" : "uploadFile",
        "parameters" : [ {
          "description" : "Vendor software product Id",
          "in" : "path",
          "name" : "vspId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor software product version Id",
          "in" : "path",
          "name" : "versionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor software product component Id",
          "in" : "path",
          "name" : "componentId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor software product process Id",
          "in" : "path",
          "name" : "processId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "USER_ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "$ref" : "#/components/schemas/Attachment"
              }
            }
          }
        },
        "responses" : {
          "default" : {
            "content" : {
              "application/json" : { }
            },
            "description" : "default response"
          }
        },
        "tags" : [ "SDCE-1 APIs", "Vendor Software Product Component Processes" ]
      }
    },
...
    "/v1.0/vendor-software-products/{vspId}/versions/{versionId}/components/{componentId}/uploads" : {
      "get" : {
        "description" : "Get the filenames of uploaded files by type",
        "operationId" : "list_4",
        "parameters" : [ {
          "description" : "Vendor software product Id",
          "in" : "path",
          "name" : "vspId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor software product version Id",
          "in" : "path",
          "name" : "versionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor software product component Id",
          "in" : "path",
          "name" : "componentId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "USER_ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/MonitoringUploadStatusDto"
                  }
                }
              }
            }
          }
        },
        "tags" : [ "SDCE-1 APIs", "Vendor Software Product Component Uploads" ]
      }
    },
...
    "/v1.0/vendor-software-products/{vspId}/versions/{versionId}/processes/{processId}/upload" : {
      "delete" : {
        "description" : "Delete vendor software product process uploaded file",
        "operationId" : "deleteUploadedFile_1",
        "parameters" : [ {
          "description" : "Vendor software product Id",
          "in" : "path",
          "name" : "vspId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor software product version Id",
          "in" : "path",
          "name" : "versionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor software product process Id",
          "in" : "path",
          "name" : "processId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "USER_ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/json" : { }
            },
            "description" : "default response"
          }
        },
        "tags" : [ "SDCE-1 APIs", "Vendor Software Product Processes" ]
      },
      "get" : {
        "description" : "Get vendor software product process uploaded file",
        "operationId" : "getUploadedFile_1",
        "parameters" : [ {
          "description" : "Vendor software product Id",
          "in" : "path",
          "name" : "vspId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor software product version Id",
          "in" : "path",
          "name" : "versionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor software product process Id",
          "in" : "path",
          "name" : "processId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "USER_ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/octet-stream" : { }
            },
            "description" : "default response"
          }
        },
        "tags" : [ "SDCE-1 APIs", "Vendor Software Product Processes" ]
      },
      "post" : {
        "description" : "Update vendor software product process upload",
        "operationId" : "uploadFile_1",
        "parameters" : [ {
          "description" : "Vendor software product Id",
          "in" : "path",
          "name" : "vspId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor software product version Id",
          "in" : "path",
          "name" : "versionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Vendor software product process Id",
          "in" : "path",
          "name" : "processId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "USER_ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "$ref" : "#/components/schemas/Attachment"
              }
            }
          }
        },
        "responses" : {
          "default" : {
            "content" : {
              "application/json" : { }
            },
            "description" : "default response"
          }
        },
        "tags" : [ "SDCE-1 APIs", "Vendor Software Product Processes" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Attachment" : {
        "type" : "object",
        "properties" : {
          "contentDisposition" : {
            "$ref" : "#/components/schemas/ContentDisposition"
          },
          "contentId" : {
            "type" : "string"
          },
          "contentType" : {
            "type" : "object",
            "properties" : {
              "parameters" : {
                "type" : "object",
                "additionalProperties" : {
                  "type" : "string"
                }
              },
              "subtype" : {
                "type" : "string"
              },
              "type" : {
                "type" : "string"
              },
              "wildcardSubtype" : {
                "type" : "boolean"
              },
              "wildcardType" : {
                "type" : "boolean"
              }
            }
          },
          "dataHandler" : {
            "type" : "object",
            "properties" : {
              "allCommands" : {
                "type" : "array",
                "items" : {
                  "type" : "object",
                  "properties" : {
                    "commandClass" : {
                      "type" : "string"
                    },
                    "commandName" : {
                      "type" : "string"
                    }
                  }
                }
              },
              "commandMap" : {
                "type" : "object",
                "properties" : {
                  "mimeTypes" : {
                    "type" : "array",
                    "items" : {
                      "type" : "string"
                    }
                  }
                },
                "writeOnly" : true
              },
              "content" : {
                "type" : "object"
              },
              "contentType" : {
                "type" : "string"
              },
              "dataSource" : {
                "type" : "object",
                "properties" : {
                  "contentType" : {
                    "type" : "string"
                  },
                  "inputStream" : {
                    "type" : "object"
                  },
                  "name" : {
                    "type" : "string"
                  },
                  "outputStream" : {
                    "type" : "object"
                  }
                }
              },
              "inputStream" : {
                "type" : "object"
              },
              "name" : {
                "type" : "string"
              },
              "outputStream" : {
                "type" : "object"
              },
              "preferredCommands" : {
                "type" : "array",
                "items" : {
                  "type" : "object",
                  "properties" : {
                    "commandClass" : {
                      "type" : "string"
                    },
                    "commandName" : {
                      "type" : "string"
                    }
                  }
                }
              },
              "transferDataFlavors" : {
                "type" : "array",
                "items" : {
                  "type" : "object",
                  "properties" : {
                    "defaultRepresentationClassAsString" : {
                      "type" : "string"
                    },
                    "flavorJavaFileListType" : {
                      "type" : "boolean"
                    },
                    "flavorRemoteObjectType" : {
                      "type" : "boolean"
                    },
                    "flavorSerializedObjectType" : {
                      "type" : "boolean"
                    },
                    "flavorTextType" : {
                      "type" : "boolean"
                    },
                    "humanPresentableName" : {
                      "type" : "string"
                    },
                    "mimeType" : {
                      "type" : "string"
                    },
                    "mimeTypeSerializedObject" : {
                      "type" : "boolean"
                    },
                    "primaryType" : {
                      "type" : "string"
                    },
                    "representationClassByteBuffer" : {
                      "type" : "boolean"
                    },
                    "representationClassCharBuffer" : {
                      "type" : "boolean"
                    },
                    "representationClassInputStream" : {
                      "type" : "boolean"
                    },
                    "representationClassReader" : {
                      "type" : "boolean"
                    },
                    "representationClassRemote" : {
                      "type" : "boolean"
                    },
                    "representationClassSerializable" : {
                      "type" : "boolean"
                    },
                    "subType" : {
                      "type" : "string"
                    }
                  }
                }
              }
            }
          },
          "headers" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "array",
              "items" : {
                "type" : "string"
              }
            }
          },
          "object" : {
            "type" : "object"
          }
        }
      },
...
      "ContentDisposition" : {
        "type" : "object",
        "properties" : {
          "filename" : {
            "type" : "string"
          },
          "parameters" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "type" : {
            "type" : "string"
          }
        }
      },
...
      "MonitoringUploadStatusDto" : {
        "type" : "object",
        "properties" : {
          "snmpPoll" : {
            "type" : "string"
          },
          "snmpTrap" : {
            "type" : "string"
          },
          "vesEvent" : {
            "type" : "string"
          }
        }
      },
...
    }
  }
}